% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{apply_custom_groups}
\alias{apply_custom_groups}
\title{Apply custom column groups to data}
\usage{
apply_custom_groups(dt, custom_groups)
}
\arguments{
\item{dt}{data.table}

\item{custom_groups}{List of tplyr_custom_group objects (or NULL)}
}
\value{
Modified data.table
}
\description{
Duplicates rows matching source levels with the column variable set to
the custom group name.
}
\keyword{internal}
