% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{apply_total_groups}
\alias{apply_total_groups}
\title{Apply total groups to data}
\usage{
apply_total_groups(dt, total_groups)
}
\arguments{
\item{dt}{data.table}

\item{total_groups}{List of tplyr_total_group objects (or NULL)}
}
\value{
Modified data.table
}
\description{
Duplicates all rows with the column variable set to the total group label,
creating a synthetic "Total" column level.
}
\keyword{internal}
