% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_analyze.R
\name{build_analyze_row_labels}
\alias{build_analyze_row_labels}
\title{Build row label columns for analyze layer}
\usage{
build_analyze_row_labels(combined, by_labels, by_data_vars)
}
\description{
Build row label columns for analyze layer
}
\keyword{internal}
