% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_desc.R
\name{build_desc_single}
\alias{build_desc_single}
\title{Build a single-target desc layer}
\usage{
build_desc_single(
  dt,
  tv,
  cols,
  by_data_vars,
  by_labels,
  settings,
  layer_index,
  col_n,
  var_label = NULL,
  var_index = NULL,
  pop_dt = NULL,
  col_levels = NULL
)
}
\description{
Build a single-target desc layer
}
\keyword{internal}
