% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_count.R
\name{build_nested_row_labels}
\alias{build_nested_row_labels}
\title{Build row labels for a nested count level}
\usage{
build_nested_row_labels(
  counts,
  by_labels,
  by_data_vars,
  target_vars,
  level,
  n_levels,
  indentation
)
}
\description{
Build row labels for a nested count level
}
\keyword{internal}
