% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_count.R
\name{build_nested_row_labels_special}
\alias{build_nested_row_labels_special}
\title{Build row labels for special rows (total/missing) in nested context}
\usage{
build_nested_row_labels_special(
  dt,
  by_labels,
  by_data_vars,
  target_vars,
  tv,
  n_label_cols
)
}
\description{
Build row labels for special rows (total/missing) in nested context
}
\keyword{internal}
