% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_count.R
\name{build_row_labels_count}
\alias{build_row_labels_count}
\title{Build row label columns for count layers}
\usage{
build_row_labels_count(counts, by_labels, by_data_vars, tv)
}
\description{
Build row label columns for count layers
}
\keyword{internal}
