% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_count.R
\name{build_row_labels_special}
\alias{build_row_labels_special}
\title{Build row labels for special rows (missing, total) in single-variable context}
\usage{
build_row_labels_special(dt, by_labels, by_data_vars, tv, existing_label_cols)
}
\description{
Build row labels for special rows (missing, total) in single-variable context
}
\keyword{internal}
