% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_shift.R
\name{build_shift_row_labels}
\alias{build_shift_row_labels}
\title{Build row labels for shift layer}
\usage{
build_shift_row_labels(counts, by_labels, by_data_vars, row_var)
}
\description{
Build row labels for shift layer
}
\keyword{internal}
