% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_analyze.R
\name{call_analyze_fn_by_group}
\alias{call_analyze_fn_by_group}
\title{Call analyze_fn for each group combination}
\usage{
call_analyze_fn_by_group(dt, analyze_fn, target_var, group_vars)
}
\description{
Call analyze_fn for each group combination
}
\keyword{internal}
