% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precision.R
\name{collect_precision_values}
\alias{collect_precision_values}
\title{Compute max integer width and max decimal precision from a numeric vector}
\usage{
collect_precision_values(vals)
}
\arguments{
\item{vals}{Numeric vector (should already have NA/Inf removed)}
}
\value{
list(max_int, max_dec)
}
\description{
Compute max integer width and max decimal precision from a numeric vector
}
\keyword{internal}
