% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_errors.R
\name{collect_user_fn_errors}
\alias{collect_user_fn_errors}
\title{Evaluate a build, reporting any user-code failures as one warning}
\usage{
collect_user_fn_errors(expr)
}
\arguments{
\item{expr}{Expression to evaluate}
}
\value{
The value of \code{expr}
}
\description{
Evaluate a build, reporting any user-code failures as one warning
}
\keyword{internal}
