% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_count.R
\name{complete_counts}
\alias{complete_counts}
\title{Complete count data to ensure all combinations exist}
\usage{
complete_counts(
  counts,
  dt,
  cols,
  by_data_vars,
  tv,
  limit_data_by = NULL,
  denom_group = NULL,
  col_levels = NULL
)
}
\description{
Complete count data to ensure all combinations exist
}
\keyword{internal}
