% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_count.R
\name{complete_nested_level}
\alias{complete_nested_level}
\title{Complete counts for a nested level}
\usage{
complete_nested_level(
  counts,
  dt,
  cols,
  by_data_vars,
  level_tvs,
  limit_data_by = NULL,
  denom_group = NULL,
  col_levels = NULL
)
}
\description{
Complete counts for a nested level
}
\keyword{internal}
