% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_count.R
\name{compute_missing_counts}
\alias{compute_missing_counts}
\title{Compute missing count row}
\usage{
compute_missing_counts(
  dt,
  counts,
  cols,
  by_data_vars,
  tv,
  group_vars,
  denom_group,
  denom_dt,
  distinct_by,
  missing_count
)
}
\description{
Compute missing count row
}
\keyword{internal}
