% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_count.R
\name{compute_total_row}
\alias{compute_total_row}
\title{Compute total row}
\usage{
compute_total_row(
  counts,
  dt,
  cols,
  by_data_vars,
  tv,
  total_label,
  total_missings,
  distinct_by,
  missing_count,
  denom_group,
  denom_dt
)
}
\description{
Compute total row
}
\keyword{internal}
