% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard.R
\name{create_desc_rows_from_ard}
\alias{create_desc_rows_from_ard}
\title{Create formatted rows for desc layer from pivoted ARD stats}
\usage{
create_desc_rows_from_ard(wide_stats, fmt_list, desc_group)
}
\description{
Create formatted rows for desc layer from pivoted ARD stats
}
\keyword{internal}
