% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_count.R
\name{get_count_format}
\alias{get_count_format}
\title{Get the count format string, falling back to defaults}
\usage{
get_count_format(settings)
}
\description{
Get the count format string, falling back to defaults
}
\keyword{internal}
