% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_desc.R
\name{get_desc_formats}
\alias{get_desc_formats}
\title{Get descriptive format strings, falling back to defaults}
\usage{
get_desc_formats(settings)
}
\description{
Get descriptive format strings, falling back to defaults
}
\keyword{internal}
