% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_count.R
\name{get_nested_denom_group}
\alias{get_nested_denom_group}
\title{Resolve denominator group for a nesting level}
\usage{
get_nested_denom_group(denoms_by, level, cols)
}
\description{
Resolve denominator group for a nesting level
}
\keyword{internal}
