% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast.R
\name{harmonize_and_bind}
\alias{harmonize_and_bind}
\title{Harmonize column sets across layers and row-bind}
\usage{
harmonize_and_bind(layer_results)
}
\description{
Harmonize column sets across layers and row-bind
}
\keyword{internal}
