% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{is_label}
\alias{is_label}
\title{Check if an object is a tplyr_label}
\usage{
is_label(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
Logical
}
\description{
Check if an object is a tplyr_label
}
\keyword{internal}
