% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_data.R
\name{is_pop_data}
\alias{is_pop_data}
\title{Check if an object is a tplyr_pop_data}
\usage{
is_pop_data(x)
}
\arguments{
\item{x}{An object to check}
}
\value{
Logical
}
\description{
Check if an object is a tplyr_pop_data
}
\examples{
is_pop_data(pop_data(cols = "TRT01P"))
is_pop_data("TRT01P")

}
