% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{make_in_filter}
\alias{make_in_filter}
\title{Build an inclusion filter expression}
\usage{
make_in_filter(var_name, values)
}
\arguments{
\item{var_name}{Character name of the variable}

\item{values}{Vector of values for the inclusion set}
}
\description{
Build an inclusion filter expression
}
\keyword{internal}
