% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{make_missing_filter}
\alias{make_missing_filter}
\title{Build a filter for missing values (is.na OR \%in\% missing_values)}
\usage{
make_missing_filter(var_name, missing_values = character(0))
}
\description{
Build a filter for missing values (is.na OR \%in\% missing_values)
}
\keyword{internal}
