% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{make_not_in_filter}
\alias{make_not_in_filter}
\title{Build an exclusion filter expression}
\usage{
make_not_in_filter(var_name, values)
}
\arguments{
\item{var_name}{Character name of the variable}

\item{values}{Vector of values to exclude}
}
\description{
Build an exclusion filter expression
}
\keyword{internal}
