% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_str.R
\name{parse_format_part}
\alias{parse_format_part}
\title{Parse one side (int or dec) of a format group}
\usage{
parse_format_part(part)
}
\description{
Parse one side (int or dec) of a format group
}
\keyword{internal}
