% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_str.R
\name{parse_format_string}
\alias{parse_format_string}
\title{Parse a format string into groups and literals}
\usage{
parse_format_string(fmt)
}
\description{
Parse a format string into groups and literals
}
\keyword{internal}
