% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard.R
\name{reconstruct_layer_from_ard}
\alias{reconstruct_layer_from_ard}
\title{Reconstruct a single layer from ARD data}
\usage{
reconstruct_layer_from_ard(layer_ard, layer, cols, layer_index)
}
\description{
Reconstruct a single layer from ARD data
}
\keyword{internal}
