% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordering.R
\name{rename_ord_columns}
\alias{rename_ord_columns}
\title{Rename ordering columns to the public output names}
\usage{
rename_ord_columns(result)
}
\arguments{
\item{result}{data.table}
}
\value{
Modified data.table (by reference)
}
\description{
Renames \code{ordindx} to \code{ord_layer_index} and \code{ord1}/\code{ord2}/... to
\code{ord_layer_1}/\code{ord_layer_2}/...
}
\keyword{internal}
