% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{serializable_to_spec}
\alias{serializable_to_spec}
\title{Convert raw parsed JSON/YAML back to tplyr_spec}
\usage{
serializable_to_spec(raw)
}
\description{
Convert raw parsed JSON/YAML back to tplyr_spec
}
\keyword{internal}
