% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tplyr_adae}
\alias{tplyr_adae}
\title{Adverse events analysis dataset}
\format{
A data.frame
}
\usage{
tplyr_adae
}
\description{
A sample CDISC ADaM ADAE dataset from the PHUSE Test Data Factory.
Contains adverse event records.
}
\keyword{datasets}
