% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tplyr_adlb}
\alias{tplyr_adlb}
\title{Laboratory data analysis dataset}
\format{
A data.frame
}
\usage{
tplyr_adlb
}
\description{
A sample CDISC ADaM ADLB dataset from the PHUSE Test Data Factory.
Contains laboratory test results.
}
\keyword{datasets}
