% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tplyr_adsl}
\alias{tplyr_adsl}
\title{Subject-level analysis dataset}
\format{
A data.frame
}
\usage{
tplyr_adsl
}
\description{
A sample CDISC ADaM ADSL dataset from the PHUSE Test Data Factory.
Contains subject-level demographics and baseline characteristics.
}
\keyword{datasets}
