% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{tplyr_layers}
\alias{tplyr_layers}
\title{Create a list of layers}
\usage{
tplyr_layers(...)
}
\arguments{
\item{...}{Layer objects created by group_count(), group_desc(),
group_shift(), or group_analyze()}
}
\value{
A list of tplyr_layer objects
}
\description{
Wraps one or more layer objects into a validated list for use in tplyr_spec().
}
\examples{
# Layers stack in the order given, and may mix types freely
layers <- tplyr_layers(
  group_desc("AGE"),
  group_count("SEX"),
  group_count("AGEGR1")
)
length(layers)

spec <- tplyr_spec(cols = "TRT01P", layers = layers)
tplyr_build(spec, tplyr_adsl)

}
