% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{tplyr_meta_anti_join}
\alias{tplyr_meta_anti_join}
\title{Anti-join metadata for missing subjects}
\usage{
tplyr_meta_anti_join(join_meta, on)
}
\arguments{
\item{join_meta}{A \code{tplyr_meta} object with filters for the population data}

\item{on}{Character vector of join key variable names (e.g., \code{"USUBJID"})}
}
\value{
A tplyr_meta_anti_join object
}
\description{
Anti-join metadata for missing subjects
}
\keyword{internal}
