% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_format_strings}
\alias{validate_format_strings}
\title{Validate format strings in layer settings}
\usage{
validate_format_strings(fmt_list, layer_index)
}
\arguments{
\item{fmt_list}{A named list expected to contain f_str objects}

\item{layer_index}{Integer layer index (for error messages)}
}
\value{
Invisible TRUE if valid
}
\description{
Validate format strings in layer settings
}
\keyword{internal}
