% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_stat_columns}
\alias{validate_stat_columns}
\title{Validate stat_columns in layer settings}
\usage{
validate_stat_columns(stat_cols, layer_index)
}
\arguments{
\item{stat_cols}{A named list expected to contain f_str objects}

\item{layer_index}{Integer layer index (for error messages)}
}
\value{
Invisible TRUE if valid
}
\description{
Validate stat_columns in layer settings
}
\keyword{internal}
