% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_count.R
\name{build_count_layer_nested}
\alias{build_count_layer_nested}
\title{Process a nested (multi-variable) count layer}
\usage{
build_count_layer_nested(
  dt,
  target_vars,
  cols,
  by_data_vars,
  by_labels,
  settings,
  layer_index,
  col_n = NULL,
  pop_dt = NULL
)
}
\description{
Process a nested (multi-variable) count layer
}
\keyword{internal}
