% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_count.R
\name{build_count_layer_single}
\alias{build_count_layer_single}
\title{Process a single-variable count layer}
\usage{
build_count_layer_single(
  dt,
  tv,
  cols,
  by_data_vars,
  by_labels,
  settings,
  layer_index,
  col_n = NULL,
  pop_dt = NULL
)
}
\description{
Process a single-variable count layer
}
\keyword{internal}
