% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_desc.R
\name{build_desc_layer}
\alias{build_desc_layer}
\title{Process a descriptive statistics layer}
\usage{
build_desc_layer(dt, layer, cols, layer_index, col_n = NULL, pop_dt = NULL)
}
\arguments{
\item{dt}{data.table with the (filtered) input data}

\item{layer}{A tplyr_desc_layer object}

\item{cols}{Character vector of column variable names from the spec}

\item{layer_index}{Integer index of this layer}
}
\value{
A data.table with rowlabel*, res*, and ord* columns
}
\description{
Process a descriptive statistics layer
}
\keyword{internal}
