% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_desc.R
\name{build_desc_multi}
\alias{build_desc_multi}
\title{Build a multi-target desc layer}
\usage{
build_desc_multi(
  dt,
  target_vars,
  cols,
  by_data_vars,
  by_labels,
  settings,
  layer_index,
  col_n,
  pop_dt = NULL
)
}
\description{
Build a multi-target desc layer
}
\keyword{internal}
