% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_count.R
\name{cast_to_wide}
\alias{cast_to_wide}
\title{Cast long data to wide output format}
\usage{
cast_to_wide(dt, row_label_cols, cols, layer_index)
}
\description{
Cast long data to wide output format
}
\keyword{internal}
