% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_count.R
\name{classify_by}
\alias{classify_by}
\title{Classify by values into data variables and labels}
\usage{
classify_by(by, col_names)
}
\description{
Classify by values into data variables and labels
}
\keyword{internal}
