% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process.R
\name{collapse_row_labels}
\alias{collapse_row_labels}
\title{Collapse row labels into a single column}
\usage{
collapse_row_labels(x, ..., indent = "  ", target_col = "row_label")
}
\arguments{
\item{x}{Input data frame}

\item{...}{Column names (as character strings) to be collapsed, must be 2 or
more}

\item{indent}{Indentation string to be used, which is multiplied at each
indentation level}

\item{target_col}{Character string naming the output column containing
collapsed row labels}
}
\value{
data.frame with row labels collapsed into a single column
}
\description{
This is a generalized post processing function that allows you to take groups
of by variables and collapse them into a single column. Repeating values are
split into separate rows, and for each level of nesting, a specified
indentation level can be applied.
}
\examples{
x <- data.frame(
  row_label1 = c("A", "A", "A", "B", "B"),
  row_label2 = c("C", "C", "D", "E", "F"),
  var1 = 1:5,
  stringsAsFactors = FALSE
)

collapse_row_labels(x, "row_label1", "row_label2")

collapse_row_labels(x, "row_label1", "row_label2", indent = "    ",
                    target_col = "rl")

}
