% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_count.R
\name{complete_counts}
\alias{complete_counts}
\title{Complete count data to ensure all combinations exist}
\usage{
complete_counts(counts, dt, cols, by_data_vars, tv)
}
\description{
Complete count data to ensure all combinations exist
}
\keyword{internal}
