% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_shift.R
\name{complete_shift_counts}
\alias{complete_shift_counts}
\title{Complete shift count data}
\usage{
complete_shift_counts(
  counts,
  dt,
  all_cols,
  by_data_vars,
  row_var,
  denom_group = NULL
)
}
\description{
Ensures all row_var × col_var × cols combinations exist, filling missing
combinations with zero counts.
}
\keyword{internal}
