% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_str.R
\name{f_str}
\alias{f_str}
\title{Create a format string object}
\usage{
f_str(format_string, ..., empty = NULL)
}
\arguments{
\item{format_string}{Character string defining the display template}

\item{...}{Character strings naming the variables that populate the template}

\item{empty}{Value(s) to display when data is NA/missing}
}
\value{
A tplyr_f_str object
}
\description{
Create a format string object
}
