% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{group_desc}
\alias{group_desc}
\title{Create a descriptive statistics layer}
\usage{
group_desc(target_var, by = NULL, where = NULL, settings = layer_settings())
}
\arguments{
\item{target_var}{Character string or vector naming the target variable(s)}

\item{by}{Character string or vector for row grouping}

\item{where}{Expression for filtering data for this layer}

\item{settings}{A layer_settings object}
}
\value{
A tplyr_desc_layer object
}
\description{
Create a descriptive statistics layer
}
