% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{group_shift}
\alias{group_shift}
\title{Create a shift layer}
\usage{
group_shift(target_var, by = NULL, where = NULL, settings = layer_settings())
}
\arguments{
\item{target_var}{Named character vector with \code{row} and \code{column} elements}

\item{by}{Character string or vector for row grouping}

\item{where}{Expression for filtering data for this layer}

\item{settings}{A layer_settings object}
}
\value{
A tplyr_shift_layer object
}
\description{
Create a shift layer
}
