% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{is_tplyr_layer}
\alias{is_tplyr_layer}
\title{Check if an object is a tplyr_layer}
\usage{
is_tplyr_layer(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
Logical
}
\description{
Check if an object is a tplyr_layer
}
