% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{is_tplyr_spec}
\alias{is_tplyr_spec}
\title{Check if an object is a tplyr_spec}
\usage{
is_tplyr_spec(x)
}
\arguments{
\item{x}{An object to check}
}
\value{
Logical
}
\description{
Check if an object is a tplyr_spec
}
