% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process.R
\name{replace_leading_whitespace}
\alias{replace_leading_whitespace}
\title{Replace leading whitespace with a specified string}
\usage{
replace_leading_whitespace(x, replace_with = " ")
}
\arguments{
\item{x}{Character vector}

\item{replace_with}{Replacement string for each leading space}
}
\value{
Character vector with leading spaces replaced
}
\description{
Useful for HTML rendering where leading spaces are collapsed.
}
