% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_count.R
\name{sort_nested}
\alias{sort_nested}
\title{Sort nested count data for correct interleaving}
\usage{
sort_nested(
  combined,
  target_vars,
  by_data_vars,
  dt = NULL,
  outer_sort_position = NULL
)
}
\description{
Sort nested count data for correct interleaving
}
\keyword{internal}
