% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process.R
\name{str_extract_num}
\alias{str_extract_num}
\title{Extract numeric values from formatted strings}
\usage{
str_extract_num(x, index = 1L)
}
\arguments{
\item{x}{Character vector of formatted strings}

\item{index}{Integer, which numeric value to extract (1-based)}
}
\value{
Numeric vector
}
\description{
Extracts the Nth numeric value from a formatted tplyr2 string.
}
