% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{tplyr_layers}
\alias{tplyr_layers}
\title{Create a list of layers}
\usage{
tplyr_layers(...)
}
\arguments{
\item{...}{Layer objects created by group_count(), group_desc(), or group_shift()}
}
\value{
A list of tplyr_layer objects
}
\description{
Wraps one or more layer objects into a validated list for use in tplyr_spec().
}
