% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{tplyr_read_spec}
\alias{tplyr_read_spec}
\title{Read a tplyr_spec from JSON or YAML}
\usage{
tplyr_read_spec(path)
}
\arguments{
\item{path}{File path to a JSON or YAML spec file}
}
\value{
A tplyr_spec object
}
\description{
Deserializes a spec from a file. Expressions are reconstructed from their
string representations.
}
