% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_desc.R
\name{transpose_stats_to_columns}
\alias{transpose_stats_to_columns}
\title{Transpose stats-as-columns}
\usage{
transpose_stats_to_columns(wide)
}
\arguments{
\item{wide}{data.table from standard desc processing}
}
\value{
Transposed data.table
}
\description{
Transposes the standard wide output so that treatment groups become rows
and stat names become columns.
}
\keyword{internal}
