% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_layer}
\alias{validate_layer}
\title{Validate a single layer}
\usage{
validate_layer(layer, index)
}
\arguments{
\item{layer}{A tplyr_layer object}

\item{index}{Integer layer index (for error messages)}
}
\value{
Invisible TRUE if valid
}
\description{
Validate a single layer
}
\keyword{internal}
