% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_count.R
\name{zero_fill_stats}
\alias{zero_fill_stats}
\title{Zero-fill count statistics left NA by grid completion or merges}
\usage{
zero_fill_stats(dt)
}
\description{
Zero-fill count statistics left NA by grid completion or merges
}
\keyword{internal}
